Remove unnecessary apps from INSTALLED_APPS.
Modules only need to be in INSTALLED_APPS if they provide models,
templates, template tags, or other such things. We have a few entries
listed that do not need to be there.

One of these, djblets.webapi, introduced problems when we added models
to reviewboard.webapi. App labels must be unique in INSTALLED_APPS,
due to several assumptions in Django, which was the source of this
problem. (Note that we still violate this all over the place, but it
doesn't seem to matter too much right now for the others.)

Reviewed at https://reviews.reviewboard.org/r/6085/