Fix visual issues with saving fields during publish.
When publishing a draft review request, we go through and save all open
inlineEditors. This was using the "save" method instead of "submit", which
caused a regression in multi-line fields that use CodeMirror. The result of
this is that the fields would disappear entirely. Switching to "submit" makes
it show the rendered version properly.

Testing done:
Made some changes to "description" and "testing done", and then tried to
publish my draft. Saw that the fields didn't disappear.

Reviewed at https://reviews.reviewboard.org/r/6100/