Fix args to error callback when publishing a draft review request.
It appears, from my testing and from the Firefox debugger, that the error
callback in DraftReviewResource.publish() has the same args as .save(), that
is, (model, xhr, ...).  I believe an earlier change to just (xhr) for the
publish() error callback was in error.

Testing Done:
Manually generated a PublishError via an extension.  Confirmed that the message
did not appear in the alert dialog without the fix, and that it did with the
fix.

Reviewed at https://reviews.reviewboard.org/r/6108/