Don't assume that profiles exist in datagrid code.
If a new user was created but the associated profile was not, loading the
dashboard would cause an error (because in 1.7.x, `User.get_profile()` won't
create the profile). This is already fixed in the 2.0.x series, because we've
completely replaced `User.get_profile()` there.

There are numerious other places where the profile is assumed to exist, but
this is the only one that we've encountered in production.

Testing done:
Ran unit tests.

Fixes bug 3083.

Reviewed at https://reviews.reviewboard.org/r/6111/