Add parent diff support to the pre-commit section of "New Review Request".
In general, the new "New Review Request" page works very well for people. One
feature that was missed by a small number of users was the ability to specify a
parent diff. This change adds back that feature.

If the given diff fails due to missing files or revisions, the user is shown
the error, but additionally given the option of uploading a parent diff. This
keeps the streamlined workflow for the general case but makes it possible for
the complex one.

While I was in here, I also plumbed through errors when creating the review
request or diff to display. I don't think we've ever hit these errors in
practice, due to the diff validation step (I had to artificially hack the code
to trigger them).

Testing done:
- Uploaded a plain diff via both DnD and the select button and saw it create
  the review request.
- Uploaded a diff which required a parent diff and saw the expected error
  message with new form.
- Uploaded diff + parent diff using both DnD and the select button, and saw it
  create the review request.
- Uploaded a broken diff and saw the right error.
- Uploaded a diff and parent diff which still required another parent to
  completely resolve all changes and saw an error.
- Checked that "Start over" worked on both error pages.
- Ran jshint
- Ran js-tests

Reviewed at https://reviews.reviewboard.org/r/6117/