Fix several issues with the original/patched file resources.
The original and patched file resources had some very busted queries
for determining which resource to use for the actual FileDiff lookup.
They were considering all DiffSets in the database, and not the ones
associated with the given review requests or draft.

When fixing this, I decided the best and cleanest way to ensure proper
results would be to split the classes into published and draft
resources. We now have a base class for the original/patched resources,
non-draft versions, and draft versions.

I also realized that no permission checking was taking place. This means
that, if a user knows/guesses the proper database IDs, they can fetch
the original or patched file from a private review request that they
don't have access to. Fixing this was as simple as adding some
permission checks.

Unit tests have been added for all of this.

Reviewed at https://reviews.reviewboard.org/r/6124/