Fix "No Summary" in dashboard with draft summaries.
The code in the dashboard that put in the italic "No Summary" for review
requests that had blank summaries wasn't ordered quite right, which meant that
if there was a draft with an empty summary present, the summary field would be
empty, but if it was merely a non-public review request, the summary field
would say "No Summary".

This change re-orders things to always override the displayed summary, whether
or not there's a draft.

Testing done:
Saw that my devserver with a bunch of draft/un-published review requests no
longer showed two different things based on the presence of a draft.

Reviewed at https://reviews.reviewboard.org/r/6136/