Show errors when trying to discard drafts.
The "Discard Draft" action would fail silently if something went wrong (for
example, if the server was inaccessible or the user was no longer logged in).
This change does some trivial plumbing to show the user an error in this case
(in the same way that we show other highly unexpected errors).

Testing done:
Loaded the page, killed my devserver, and tried to discard a draft. Saw an
alert that said "HTTP 0 error", which is at least better than nothing.

Sort-of addresses bug 3453.

Reviewed at https://reviews.reviewboard.org/r/6137/