Include the revision in the "New Review Request" post-commit view.
This change adds the revision string to the list of commits in the post-commit
section of the "New Review Request" page. If the revision looks like a SHA-1
hash, we only show the first seven characters. Otherwise, we show the full
thing.

This also localizes the text in this view (I'm not sure how the "by ..."
slipped through).

Testing done:
Connected a github repo and opened the New Review Request page. Saw commit IDs.

Fixes bug 3402.

Reviewed at https://reviews.reviewboard.org/r/6140/