Update to the latest less.js
When running the devserver, I noticed that I wasn't getting retina icons
anywhere. The issue here is that the version of the less.js that we loaded when
running with `DEBUG=True` had a bug when importing mixins using `(reference)`.
I've updated our shipped copy to the latest release (1.7.3) and it fixes the
issue.

Testing done:
Loaded a review request and the admin dashboard and saw retina icons
everywhere.

Reviewed at https://reviews.reviewboard.org/r/6152/