Fix setting the height of the review request box for long fields.
The fixes to set the proper widths of the fields broke the heights.
It removed the defer() call and replaced it with a throttle(). The
defer() is important to the calculations, though, since we schedule a
layout resize after closing CodeMirror fields but before all layout has
settled.

This restores the original _scheduleResizeLayout() function and adds the
throttle there.

Reviewed at https://reviews.reviewboard.org/r/6154/