Add documentation on API tokens.
This adds documentation on authenticating with API tokens and writing
custom API policies.

The authentication docs now cover how to log in with either API tokens
or a username/password. Some of the existing docs here are also fleshed
out to create better examples.

The new API token policy docs cover the structure of policies, the
various sections of the policy, how they affect access, and some tips.
This will complement the UI for editing tokens, which is currently in
development.

All API resource pages now also contain the policy IDs for use in the
policies.

Reviewed at https://reviews.reviewboard.org/r/6155/