Upgrade to CodeMirror 3.23.
This release fixes some issues with Markdown and some issues with
editing JSON payloads (needed for the upcoming API token policy editor).

This is not the latest CodeMirror, but we had issues with the 4.x series
(4.2, specifically) last time we tried it. There's a 3.24 release, but
it's not on the online minifier yet.

The build of CodeMirror is also updated to support JSON/JavaScript
modes, and to support optional linting and brace matching.

Reviewed at https://reviews.reviewboard.org/r/6156/