Added validation for API policies.
This introduces WebAPIToken.validate_policy, which will look through a
policy definition and try to find any errors. It will check for required
data, data in invalid formats, policy IDs validity, and more.

This is called when posting or updating a policy, and the error will be
returned in the API error.

Reviewed at https://reviews.reviewboard.org/r/6164/