Pass SVN credentials when accepting a self-signed certificate.
When accepting a self-signed certificate, the SVN credentials weren't
being passed. This could result in a breakage for installs that blocked
anonymous access.

accept_certificate() is now set up to take the credentials, which the
form passes. Those are then passed to build_client for SVNTool.

This fixed some issues I was able to hit with adding repositories with a
self-signed cert and that required authentication.

While here, I also fixed a missing space between two sentences in an
error message that I came across when setting up an environment for
this.

Reviewed at https://reviews.reviewboard.org/r/6171/