Move the common inline editor CSS rules into common.less.
The inline editor CSS rules were in reviews.less, which wasn't very
useful to pages other than the review pages. The new inlineEditor usage
in the My Account page didn't have the proper styling because of this.

This change moves those rules into common.less. It also changes the
button input margins to be on the left instead of the right, and to
remove the left margin of the first input. (It's easier to remove
leading spacing on the first element than the last, since :first-child
is better supported.)

The styles specific to editors on the Review Request box have also
been moved into the .review-request, in order to prevent them from
affecting other editors unintentionally.

Reviewed at https://reviews.reviewboard.org/r/6174/