Add buttons for the file attachment caption editor.
The file attachment caption editor didn't show any buttons, which was
confusing to users. The problem was that there wasn't a great place to
put the buttons.

This change fixes this by adding a section just below the editor. The
buttons appear on a greyish background, and sort of just appear below
the thumbnail, overlapping any thumbnails that may be below it. The
reason it overlaps it to prevent any shifting/staggering of other
thumbnails.

Some work had to be done to properly set offsets and to change sizes of
the caption elements, in order to cleanly overlap anything below.
The sizes are set so that the overlap completely obscures the header
text from a thumbnail below, which is nicer-looking than ony
partially-obscuring the text.

Reviewed at https://reviews.reviewboard.org/r/6176/