Fixed extra whitespace issues in review request editors.
The file attachment thumbnails and editors on the right-hand side
of review requests had bugs where they would show extra leading
whitespace on edit. This happened because the HTML had whitespace for
formatting, which ended up in the edit field.

The templates for these fields are now compressed down to not have any
extra whitespace. This makes the templates less readable, but ensures
that the only whitespace in there is whitespace provided by the value
server-side.

Reviewed at https://reviews.reviewboard.org/r/6178/