Fix original-file and patched-file tests.
Some of the tests for the original-file and patched-file resources were failing
because it was trying to decode the responses as JSON. The problem was that one
of the GET test implementations was updated for the new
`basic_get_returns_json` field but not another. I've made an equivalent change
to handle it.

Testing done: Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/6211/