Fix usage of _api_get in GitLab.
A change to the GitLab implementation to add support for more than 100 repos
ended up breaking things, because it changed the return signature of _api_get
without fixing up all the call sites. With the addition of the headers, we now
need to extract the first part of the tuple instead of just using the return
value directly.

Testing done: Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/6210/