Improve progress reporting in condensediffs.
condensediffs now shows more information on the progress of the
conversion of diffs. Along with the existing percentage, it will also
now show the number of processed items, how many there are total, and
the estimated time to completion.

Since the estimated time to completion is based on the average time
taken per item, and the number of items, the estimate becomes more
accurate over time. Instead of showing some useless numbers up-front, we
wait 30 seconds until we show the estimate. From my testing, this
appears to be a a good number, as the estimate settled around that point
in the databases I used.

This helps give administrators an idea as to how long the operation will
take, which is useful for knowing whether to keep it going or to abort
and finish later.