Fix posting changes with an empty commit ID.
Posting with an empty commit ID (a blank string) resulted in the blank
string being set for the review request or draft. This resulted in
conflicts, since a blank string is considered a valid value for unique
constraints. We now normalize it to None in this case.

Reviewed at https://reviews.reviewboard.org/r/6219/