Fix the revision selector on IE 9/10/11.
The code for the revision selector used a property for the page scroll
offset that wasn't standard across all browsers. That property is just
an alias for the actual standard property, pageXOffset. Switching to
this fixes the revision slider on IE.

Fixes bug #3373
Reviewed at https://reviews.reviewboard.org/r/6230/