Add Webhooks and Hosting Accounts to the admin sidebar.
This adds handy links for managing or adding Webhooks or hosting
accounts to the admin UI's sidebar. These are some of the more common
database tables that may need to be modified, so we don't want users to
have to dig for them.

While in here, I cleaned up some indentation issues.

Reviewed at https://reviews.reviewboard.org/r/6259/