Fix accepting certificates for perforce repositories.
When we fixed accepting certificates for subversion repositories that used a
username/password, we broke accepting certificates for perforce. The issue here
is that the perforce implementation of accept_certificate() didn't have
username and password parameters, and so passing them in caused it to
TypeError.

Testing done:
- Set up an SSL p4 server with a self-signed certificate and saw that I could
  once again add it to Review Board.
- Ran unit tests.

Fixes bug 3512.

Reviewed at https://reviews.reviewboard.org/r/6273/