Remove the commit ID on a review request when discarded.
When posting a change for review with a commit ID and then discarding
it, that discarded review request still retains the commit ID. This
results in later conflicts with newly posted review requests, which
don't appear to break until the draft is published.

We now always remove the commit ID from the review request in order to
prevent any conflicts. While this is removing some history, it's history
that causes more harm than good and, since the review request is
discarded, it's not likely that the commit ID association is needed.