Smoothly show the dashboard on page load.
The dashboard auto-sizes to fit after load, and this was very noticeable
and ugly. We now hide the whole dashboard container at first, and then
show it just before the resize. This causes it to appear and be sized
correctly all at once, making for a much nicer experience.

Reviewed at https://reviews.reviewboard.org/r/6271/