Fix interdiffs when comparing revisions with stripped-slash filenames.
The leading-slash stripping code strikes again. In this case, an initial
revision of a diff was uploaded prior to an upgrade, and a new version was
uploaded subsequent. The result is that the filenames don't match exactly
because one has leading slashes and the other doesn't.

This change normalizes the filenames to always compare them without leading
slashes when figuring out which files match up with which.

Testing done:
Created two diff revisions and changed the FileDiffs in the database so that
the first one had leading slashes and the second didn't. Before this change, I
see two entries in the interdiff for that file. After, it shows the correct
interdiff.

Tested all permutations of leading slash/no leading slash in interdiff
combinations.

Reviewed at https://reviews.reviewboard.org/r/6272/