Fix a handful of pyflakes errors.
A few errors have snuck in over the past couple months. In the admin widgets,
the widget_select method had a couple unused variables. In the diffviewer
tests, a merge caused us to have two copies of the FileDiffMigrationTests (I've
kept the longer one). Finally, in the extensions tests, we were assigning a
hook to a variable which we didn't need.

Testing done:
- Ran pyflakes.
- Ran unit tests.