Don't use <br/> tags for newlines in Markdown text in e-mails.
We had the "nl2br" filter applied for Markdwon text, which caused every
newline to have a <br/>. However, these appeared inside a <pre/>, and we
still had newlines. The end result was that we had a blank line for
every line of text.

I can't see any reason why we'd need this filter, so I've removed it.
This results in proper-looking e-mails.

Reviewed at https://reviews.reviewboard.org/r/6286/