Fix showing Markdown field changes with 'None' values.
When rendering a Markdown field in a change description box, we weren't
ensuring that the old or new values were strings. They could be None,
which is a valid value indicating there was no content before. This
would end up breaking render_markdown().

We now convert to an empty string if it's None.

Fixes bug #3549
Reviewed at https://reviews.reviewboard.org/r/6290/